fix: Avoid endless upload loop
If the UploadNotionBlocks function encounters an error for one or more
blocks, it will not delete the block(s) from the database. This causes
the state machine to think there is work left to be done and so an
endless loop is entered.

Prevent this by having the function indicate whether all of its assigned
blocks were uploaded successfully or not. In the case of unsuccessful
blocks, the state machine will gracefully end.

Additionally, since an error in even a single block means that the
Notion page would be incomplete, the function will now fail fast rather
than continuing with the rest of the blocks in the batch. Even with a
single failed block, the page needs to be thrown away and tried again
after diagnosing what went wrong.