feat(upload_blocks): Speed up parent page identification
There is no need to download the parenturl.txt file for every block.
Just download it once and reuse it. This should marginally speed up the
function's block pace, but I haven't collected enough data to quantify
this.

This does have the side effective of the tmp dir being cleaned up only
after the entire set of blocks is processed, as opposed to the prior
behavior of after every block. I don't expect this to be an issue as it
still allows ~500MB of embedded content per set of 50 blocks.