fix(upload_blocks): strip whitespace from parent url
A byproduct of `readline()` is that it appends the end of line marker
('\n' for text files) to the returned string, even if the file doesn't
contain a new line. Notion throws an error when looking up the parent
page when the given page ID has a newline in it.

This commit strips the newline from the returned string.